Migrated db modules, removed obsolete subs file
Db functions don't have typed returns yet, this will be done next
