+++ src/shard/commands/lang.ts
@@ -1,5 +1,5 @@
 import { CmdFn } from ".";
-import { getLang, setLang } from "../../subs";
+import { getLang, setLang } from "../../db/guilds";
 import { formatLanguages, FORMAT_POST_EMBEDS } from "../format";
 import { supportedLangs } from '../../../config.json';
 import { embeds, translated } from "../post";
